Modularize app architecture
* Introduce util
* Move Logger to logger
* Move client-related functions to client
